fix: bookmark button working as intended as well as some minor code refactoring
