+++ app/src/main/java/dev/iamwami/app/quotinews/ui/splashscreens/SplashScreen.kt
@@ -1,6 +1,5 @@
 package dev.iamwami.app.quotinews.ui.splashscreens
 
-import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -18,7 +17,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import dev.iamwami.app.quotinews.R
-import dev.iamwami.app.quotinews.ui.home.HomeViewModel
 import dev.iamwami.app.quotinews.ui.util.HomeScreen
 import dev.iamwami.app.quotinews.util.Fonts
 import kotlinx.coroutines.delay
@@ -34,7 +32,6 @@ fun NavigateToSplashScreen(navController: NavHostController) {
         splashScreenState = true
 
         delay(2000)
-        Log.d("testing", "$splashScreenState")
         if (splashScreenState) {
             navController.navigate(HomeScreen.route)
         }
